Backend/requirements.txt
fastapi
uvicorn
motor
pytesseract
Pillow
pdf2image
httpx
python-dotenv
flask
motor